refactor: reorganize source by MCP, Notion, and tool domains
