+++ src/tools/content/markdown-input.ts
@@ -1,0 +1,1 @@
+export { parseMarkdownToSimpleContent } from "./simple-content.js";
