--- src/types/index.ts
@@ -1,11 +1,0 @@
-/**
- * Type definitions for Notion API
- *
- * This file re-exports all types from more specialized files
- */
-
-export * from "./args.js";
-// Export all types
-export * from "./common.js";
-export * from "./responses.js";
-export * from "./schemas.js";
